after DB refactoring

+++ src/main/java/pjatk/socialeventorganizer/SocialEventOrganizer/service/CateringService.java
@@ -7,6 +7,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import pjatk.socialeventorganizer.SocialEventOrganizer.mapper.CateringMapper;
 import pjatk.socialeventorganizer.SocialEventOrganizer.model.dto.Catering;
+import pjatk.socialeventorganizer.SocialEventOrganizer.model.exception.IllegalArgumentException;
 import pjatk.socialeventorganizer.SocialEventOrganizer.model.exception.NotFoundException;
 import pjatk.socialeventorganizer.SocialEventOrganizer.model.request.CateringRequest;
 import pjatk.socialeventorganizer.SocialEventOrganizer.model.response.CateringResponse;
@@ -30,31 +31,38 @@ public class CateringService {
         return ImmutableList.copyOf(cateringList);
     }
 
-    public Catering findByName(String name) {
-        return repository.findByName(name)
-                .orElseThrow(() -> new NotFoundException("Catering with name " + name + " not found"));
-    }
-
-    public ImmutableList<Catering> findByCity(String city) {
-        final List<Catering> cateringList = repository.findByCity(city);
+    public ImmutableList<Catering> findByName(String name) {
+        final List<Catering> cateringList = repository.findByNameContaining(name);
         if (cateringList != null && !cateringList.isEmpty()) {
             return ImmutableList.copyOf(cateringList);
         } else {
-            throw new NotFoundException("No catering in the city " + city + " was found");
+            throw new NotFoundException("No catering with the name " + name + " was found");
         }
-
     }
+
+//    public ImmutableList<Catering> findByCity(String city) {
+//        final List<Catering> cateringList = repository.findByCityContaining(city);
+//        if (cateringList != null && !cateringList.isEmpty()) {
+//            return ImmutableList.copyOf(cateringList);
+//        } else {
+//            throw new NotFoundException("No catering in the city " + city + " was found");
+//        }
+//    }
 
     public CateringResponse addNewCatering(CateringRequest request) {
         final Catering catering = cateringMapper.mapToDTO(request);
         log.info("TRYING TO SAVE" + catering.toString());
         repository.save(catering);
-
+        log.info("SAVED");
         return cateringMapper.mapToResponse(catering);
     }
 
 
     public void updateCatering(Long cateringId, CateringRequest request) {
+        if (!cateringWithIdExists(cateringId)) {
+            throw new IllegalArgumentException("Catering with ID " + cateringId + " does not exist");
+        }
+
         final Catering catering = cateringMapper.mapToDTO(request, cateringId);
         log.info("TRYING TO UPDATE " + catering);
         repository.save(catering);
@@ -62,8 +70,17 @@ public class CateringService {
     }
 
     public void deleteCatering(Long id) {
+        if (!cateringWithIdExists(id)) {
+            throw new IllegalArgumentException("Catering with ID " + id + " does not exist");
+        }
+
         log.info("TRYING TO DELETE CATERING WITH ID " + id);
         repository.deleteById(id);
+    }
+
+    public boolean cateringWithIdExists(Long id) {
+        log.info("CHECKING IF CATERING WITH ID " + id + " EXISTS");
+        return repository.existsById(id);
     }
 
 }
